Move tests/mock_provider files up a level
In order to remove Makefile.am duplication
